refactor(importers): separate CSV parsing from sample planning
